Better email template organization
